@@ -18,16 +18,26 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-
-
+  flutter_cache_manager: ^3.0.1
+  scroll_to_index: ^2.0.0
+  charcode: ^1.2.0
+  photo_view: ^0.11.1
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.0
+  cupertino_icons: ^1.0.2
+  flutter_signin_button: ^2.0.0
+  # 高级扩展 类似把dart变成kotlin那里舒适
+  supercharged: ^2.0.0
+  flash: ^1.5.1
+  flutter_bloc: ^7.0.0
+  equatable: ^2.0.0
+  wave: ^0.2.0
+  sliver_tools: ^0.2.1
 
 dev_dependencies:
   flutter_test:
@@ -43,6 +53,9 @@ flutter:
   # included with your application, so that you can use the icons in
   # the material Icons class.
   uses-material-design: true
+  assets:
+    - assets/qdy.jpg
+  
 
   # To add assets to your application, add an assets section, like this:
   # assets:
